refactor user module + groups UI

+++ src/main/java/facebook/app/dao/UserDAO.java
@@ -1,7 +1,6 @@
 package facebook.app.dao;
 import facebook.app.entities.User;
 import facebook.app.exceptions.UserIOException;
-
 import java.io.*;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Paths;
@@ -11,10 +10,8 @@ import java.util.Objects;
 
 public class UserDAO {
     private static final String FILE_NAME = "users.txt";
-
     public List<User> readUsers() throws UserIOException {
         List<User> userList = new ArrayList<>();
-
         try (InputStream inputStream = getClass().getClassLoader().getResourceAsStream(FILE_NAME)) {
             assert inputStream != null;
             try (BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream, StandardCharsets.UTF_8))) {
@@ -22,7 +19,7 @@ public class UserDAO {
                 String line;
                 while ((line = reader.readLine()) != null) {
                     String[] userData = line.split(";");
-                    long userId = Long.parseLong(userData[0].trim());
+                    int userId = Integer.parseInt(userData[0].trim());
                     String email = userData[1].trim();
                     String password = userData[2].trim();
                     boolean isLoggedIn = Boolean.parseBoolean(userData[3].trim()); // New field
@@ -44,21 +41,17 @@ public class UserDAO {
         }
         return userList;
     }
-
     public void writeUsers(List<User> userList) {
         try (OutputStream outputStream = new FileOutputStream(Objects.requireNonNull(getFilePath()));
              BufferedWriter writer = new BufferedWriter(new OutputStreamWriter(outputStream, StandardCharsets.UTF_8))) {
-
             for (User user : userList) {
                 writer.write(user.getUserId() + ";" + user.getEmail() + ";" + user.getPassword() + ";" + user.isLoggedIn());
                 writer.newLine();
             }
-
         } catch (IOException e) {
             e.printStackTrace();
         }
     }
-
     public User getUserByID(int userId) throws UserIOException {
         List<User> userList = readUsers();
         return userList.stream()
@@ -66,7 +59,6 @@ public class UserDAO {
                 .findFirst()
                 .orElse(null);
     }
-
     public User getUserByEmail(String email) throws UserIOException {
         List<User> userList = readUsers();
         return userList.stream()
@@ -74,7 +66,6 @@ public class UserDAO {
                 .findFirst()
                 .orElse(null);
     }
-
     private String getFilePath() {
         try {
             return Paths.get(Objects.requireNonNull(getClass().getClassLoader().getResource(FILE_NAME)).toURI()).toString();
@@ -83,5 +74,4 @@ public class UserDAO {
             return null;
         }
     }
-
 }
